refactor: move calendar type to useCalendar hook file
